style: apply coding convention
