--- src/main/java/vendingmachine/Constants/CoinConstant.java
@@ -1,10 +1,0 @@
-package vendingmachine.Constants;
-
-public class CoinConstant {
-
-    public static String NON_NUMERIC_ERROR_MESSAGE = "[ERROR] 금액은 숫자로만 이루어져야 합니다.";
-    public static String DIVIDE_TEN_ERROR_MESSAGE= "[ERROR] 금액은 10으로 나누어 떨어져야 합니다.";
-
-    public static String NUMERIC_REGEX = "[0-9]+";
-
-}
