+++ src/main/java/vendingmachine/controller/CoinController.java
@@ -5,11 +5,11 @@ import vendingmachine.view.OutputView;
 
 public class CoinController {
 
-    CoinService coinService = new CoinService();
+	CoinService coinService = new CoinService();
 
-    public void generate() {
-        OutputView.printMachineMoney();
-        coinService.generate();
-        OutputView.printCoin();
-    }
+	public void generate() {
+		OutputView.printMachineMoney();
+		coinService.generate();
+		OutputView.printCoin();
+	}
 }
