+++ src/main/java/vendingmachine/controller/VendingMachineController.java
@@ -3,13 +3,13 @@ package vendingmachine.controller;
 import vendingmachine.repository.Products;
 
 public class VendingMachineController {
-    CoinController coinController = new CoinController();
-    ProductController productController = new ProductController();
-    OrderController orderController = new OrderController();
+	CoinController coinController = new CoinController();
+	ProductController productController = new ProductController();
+	OrderController orderController = new OrderController();
 
-    public void run(){
-        coinController.generate();
-        Products products = productController.generate();
-        orderController.get(products);
-    }
+	public void run() {
+		coinController.generate();
+		Products products = productController.generate();
+		orderController.get(products);
+	}
 }
