+++ src/main/java/vendingmachine/repository/Money.java
@@ -1,19 +1,18 @@
 package vendingmachine.repository;
 
 public class Money {
-    private int remainder;
+	private int remainder;
 
-    public Money(int remainder) {
-        this.remainder = remainder;
-    }
+	public Money(int remainder) {
+		this.remainder = remainder;
+	}
 
+	public int getRemainder() {
+		return this.remainder;
+	}
 
-    public int getRemainder() {
-        return this.remainder;
-    }
-
-    public void subtract(int amount){
-        remainder -= amount;
-    }
+	public void subtract(int amount) {
+		remainder -= amount;
+	}
 
 }
