+++ src/main/java/vendingmachine/repository/Product.java
@@ -1,34 +1,34 @@
 package vendingmachine.repository;
 
 public class Product {
-    private String name;
-    private int price;
-    private int stock;
+	private String name;
+	private int price;
+	private int stock;
 
-    public Product(String name, int price, int stock) {
-        this.name = name;
-        this.price = price;
-        this.stock = stock;
-    }
+	public Product(String name, int price, int stock) {
+		this.name = name;
+		this.price = price;
+		this.stock = stock;
+	}
 
-    public int getPrice() {
-        return this.price;
-    }
+	public int getPrice() {
+		return this.price;
+	}
 
-    public void subtractStock() {
-        this.stock--;
-    }
+	public void subtractStock() {
+		this.stock--;
+	}
 
-    public boolean isLowerPrice(int comparedPrice) {
-        return this.price <= comparedPrice;
-    }
+	public boolean isLowerPrice(int comparedPrice) {
+		return this.price <= comparedPrice;
+	}
 
-    public boolean isOutOfStock() {
-        return stock == 0;
-    }
+	public boolean isOutOfStock() {
+		return stock == 0;
+	}
 
-    public boolean isSameName(String input) {
-        return name.equals(input);
-    }
+	public boolean isSameName(String input) {
+		return name.equals(input);
+	}
 
 }
