+++ src/main/java/vendingmachine/service/CoinValidator.java
@@ -1,24 +1,23 @@
 package vendingmachine.service;
 
-import vendingmachine.Constants.CoinConstant;
-import vendingmachine.repository.Coin;
+import vendingmachine.constants.CoinConstant;
 
 public class CoinValidator {
 
-    public void isValid(String number) {
-        if (!isNumeric(number)) {
-            throw new IllegalArgumentException(CoinConstant.NON_NUMERIC_ERROR_MESSAGE);
-        }
-        if (!isDivideByTen(number)) {
-            throw new IllegalArgumentException(CoinConstant.DIVIDE_TEN_ERROR_MESSAGE);
-        }
-    }
+	public void isValid(String number) {
+		if (!isNumeric(number)) {
+			throw new IllegalArgumentException(CoinConstant.NON_NUMERIC_ERROR_MESSAGE);
+		}
+		if (!isDivideByTen(number)) {
+			throw new IllegalArgumentException(CoinConstant.DIVIDE_TEN_ERROR_MESSAGE);
+		}
+	}
 
-    public boolean isNumeric(String number) {
-        return number.matches(CoinConstant.NUMERIC_REGEX);
-    }
+	public boolean isNumeric(String number) {
+		return number.matches(CoinConstant.NUMERIC_REGEX);
+	}
 
-    public boolean isDivideByTen(String number) {
-        return Integer.parseInt(number) % 10 == 0;
-    }
+	public boolean isDivideByTen(String number) {
+		return Integer.parseInt(number) % 10 == 0;
+	}
 }
