+++ src/main/java/vendingmachine/view/InputView.java
@@ -4,16 +4,16 @@ import camp.nextstep.edu.missionutils.Console;
 
 public class InputView {
 
-    public static String getMoney() {
-        return Console.readLine();
-    }
+	public static String getMoney() {
+		return Console.readLine();
+	}
 
-    public static String getProduct() {
-        return Console.readLine();
-    }
+	public static String getProduct() {
+		return Console.readLine();
+	}
 
-    public static String getOrder() {
-        return Console.readLine();
-    }
+	public static String getOrder() {
+		return Console.readLine();
+	}
 
 }
